strncmp, strnstr

+++ ft_strlen.c
@@ -14,9 +14,9 @@
 //#include <stdio.h>
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
